# openapi.yml
openapi: 3.1.0
info:
  title: Tracker Server API
  description: |
    API for time tracking with three main roles: Work, Learn, and Rest.
    This service helps track time spent on various tasks and manage schedules.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development (Docker host mapping 8080->3000)
  - url: http://localhost:3000
    description: Local server (direct)
...
paths:
...
  /api/v1/manage/task/create:
    post:
      tags:
        - Management
      summary: Create a new task
      description: Create a new task with a specific role
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/manage/plan-percents:
    get:
      tags:
        - Management
      summary: Get plan percents configuration
      operationId: getPlanPercents
      responses:
        '200':
          description: Plan percents retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanPercentsResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/manage/plan-percents/{group}/{value}:
    delete:
      tags:
        - Management
      summary: Remove a plan percent value
      operationId: removePlanPercent
      parameters:
        - name: group
          in: path
          required: true
          description: Plan percent group to update
          schema:
            type: string
            enum: [plan, work, learn, rest]
        - name: value
          in: path
          required: true
          description: Percent value to remove
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Plan percent removed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        '400':
          description: Invalid group or value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Value not found in the specified group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    PlanPercentsResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: "#/components/schemas/PlanPercentsData"
      required:
        - status
        - data
      example:
        status: "success"
        data:
          title: "Procent Info"
          date: "1 January 2024"
          current_choice: 0
          plans:
            - plan
            - work
            - learn
            - rest
          plan: [4, 8, 12]
          work: [20, 25, 30]
          learn: [10, 15]
          rest: [5, 10]
    PlanPercentsData:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        current_choice:
          type: integer
        plans:
          type: array
          items:
            type: string
        plan:
          type: array
          items:
            type: integer
        work:
          type: array
          items:
            type: integer
        learn:
          type: array
          items:
            type: integer
        rest:
          type: array
          items:
            type: integer
      required:
        - title
        - date
        - current_choice
        - plans
        - plan
        - work
        - learn
        - rest
...
    CreateTaskRequest:
      type: object
      properties:
        task_name:
          type: string
          description: Name of the task
        role:
          type: string
          description: Role of the task (work, learn, rest)
      required:
        - task_name
        - role
      example:
        task_name: "Programming"
        role: "learn"
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["accept", "success"]
        message:
          type: string
      required:
        - status
      example:
        status: "accept"
        message: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: ["error"]
        message:
          type: string
      required:
        - status
        - message
      example:
        status: "error"
        message: "An error occurred"

    # Schedule schemas
...